:art: organization + styled components
