+++ public/locales/en/translations.json
@@ -42,9 +42,6 @@
   "FOIA": "FOIA",
   "Data Release -": "Data Release -",
   "dataReleaseVersion": "Unreleased",
-  "Upload Ids": "Upload Ids",
-  "Filters": "Filters",
-  "All Filters": "All Filters",
   "Export TSV": "Export TSV",
   "Query saved succesfully!": "Query saved succesfully!",
   "View in My Saved Queries": "View in My Saved Queries",
@@ -66,6 +63,9 @@
     "missingDownloadPermissions": "Your account does not have the required permission to download this file.",
     "noneSelected": "If you have not selected any files, all files in your query will be included in the actions."
   },
+  "Filters": "Filters",
+  "All Filters": "All Filters",
+  "Upload Ids": "Upload Ids",
   "Browse": "Browse",
   "Matching files in the Kids First Data Repository": "Matching files in the Kids First Data Repository",
   "Actions": "Actions",
