+++ src/components/FileRepoSidebar/FileRepoSidebar.js
@@ -13,57 +13,51 @@ import FileManifestsDownloadInput from './FileManifestsDownloadInput';
 import Subsection from './Subsection';
 import ReportsDownloadInput from './ReportsDownloadInput';
 
-const styles = ({ containerWidth, contentSidePadding, expanded, theme }) => ({
-  container: css`
-    overflow-y: auto;
-    background-color: #f4f5f8;
-    flex-grow: 0;
-    flex-shrink: 1;
-    width: ${containerWidth + contentSidePadding * 2}px;
-    min-width: 265px;
-    height: 100%;
-  `,
-  titleBar: css`
-    background-color: ${theme.greyScale5};
-    margin: 0px;
-    display: flex;
-    padding-top: 15px;
-    padding-left: 15px;
-    cursor: pointer;
-  `,
-  content: css`
-    padding-left: ${expanded ? contentSidePadding : contentSidePadding * 10}px;
-    overflow: hidden;
-    padding-right: ${contentSidePadding}px;
-    transition: all 0.25s;
-    padding-top: 10px;
-  `,
-});
+const Slideable = styled('div')`
+  position: relative;
+  transition: all 0.25s;
+  width: ${({ expanded, containerWidth, contentSidePadding }) =>
+    expanded ? `${containerWidth + contentSidePadding * 2}px` : '40px'};
+`;
+
+const Container = styled('div')`
+  overflow-y: auto;
+  box-shadow: 0 0 4.9px 0.2px ${({ theme }) => theme.shadow};
+  flex-grow: 0;
+  flex-shrink: 1;
+  width: ${({ containerWidth, contentSidePadding }) => containerWidth + contentSidePadding * 2}px;
+  min-width: 265px;
+  height: 100%;
+`;
+
+const Titlebar = styled('div')`
+  background-color: ${({ theme }) => theme.greyScale5};
+  margin: 0px;
+  display: flex;
+  padding-top: 15px;
+  padding-left: 15px;
+  cursor: pointer;
+`;
+
+const Content = styled('div')`
+  padding-left: ${({ expanded, contentSidePadding }) =>
+    expanded ? contentSidePadding : contentSidePadding * 10}px;
+  overflow: hidden;
+  padding-right: ${({ contentSidePadding }) => contentSidePadding}px;
+  transition: all 0.25s;
+  padding-top: 10px;
+`;
+
+const Text = styled('div')`
+  font-size: 14px;
+  line-height: 26px;
+`;
 
 const Divider = styled('div')`
   height: 1px;
-  background-color: #d4d6dd;
+  background-color: ${({ theme }) => theme.greyScale8};
   margin: 20px 0px;
 `;
-
-const SlidablePanel = ({ expanded, containerWidth, contentSidePadding, ...rest }) => {
-  return (
-    <div
-      className={css`
-        position: relative;
-        transition: all 0.25s;
-        width: ${expanded ? `${containerWidth + contentSidePadding * 2}px` : '40px'};
-        border-color: #c6c7cc;
-        border-style: solid;
-        border-width: 0 0 0 1px;
-        box-sizing: border-box;
-        box-shadow: 0 0 4.9px 0.2px #a0a0a3;
-        overflow: hidden;
-      `}
-      {...rest}
-    />
-  );
-};
 
 const FileRepoSidebar = compose(withTheme, withState('expanded', 'setExpanded', true))(
   ({
@@ -73,16 +67,11 @@ const FileRepoSidebar = compose(withTheme, withState('expanded', 'setExpanded', 
     style,
     containerWidth = 310,
     contentSidePadding = 15,
-    panelStyle = styles({ containerWidth, contentSidePadding, theme, expanded }),
     ...props
   }) => (
-    <SlidablePanel {...{ contentSidePadding, containerWidth, expanded }}>
-      <div
-        css={`
-          ${panelStyle.container} ${style};
-        `}
-      >
-        <div className={panelStyle.titleBar} onClick={() => setExpanded(!expanded)}>
+    <Slideable {...{ contentSidePadding, containerWidth, expanded }}>
+      <Container {...{ contentSidePadding, containerWidth }}>
+        <Titlebar onClick={() => setExpanded(!expanded)}>
           <Heading>
             <span
               className={css`
@@ -98,19 +87,14 @@ const FileRepoSidebar = compose(withTheme, withState('expanded', 'setExpanded', 
             </span>
             <Trans>Actions</Trans>
           </Heading>
-        </div>
-        <div className={panelStyle.content}>
-          <div
-            css={`
-              font-size: 14px;
-              line-height: 26px;
-            `}
-          >
+        </Titlebar>
+        <Content {...{ expanded, contentSidePadding, containerWidth }}>
+          <Text>
             <Trans i18nKey="fileRepoSidebar.noneSelected">
               If you have not selected any files, all files in your query will be included in the
               actions.
             </Trans>
-          </div>
+          </Text>
           <Divider />
           <Heading>
             <Trans>Download</Trans>
@@ -128,9 +112,9 @@ const FileRepoSidebar = compose(withTheme, withState('expanded', 'setExpanded', 
             <Trans>Data Analysis</Trans>
           </Heading>
           <CavaticaCopyButton {...props} />
-        </div>
-      </div>
-    </SlidablePanel>
+        </Content>
+      </Container>
+    </Slideable>
   ),
 );
